bc: Fixup stack top
Add precise stack top tracking, untangling next register (nr) and destinatino register (rd)
in bc, and then making sure we're not off by one in both the vm, and in the trace recorder.

The main advantage is we don't sometimes save an extra stack slot in snapshots, and store it
when we trace a loop: For example in diviter stores one less stack slot.